vote/requirements.txt
Flask>=2.0
confluent-kafka>=1.9.0
python-dotenv # Good for local testing, reads .env file if present